requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
Pillow
django-phonenumber-field